[LASB-3000] Move the means.json to the general folder.
