+++ crime-applications-adaptor/src/main/resources/schemas/criminalapplicationsdatastore/general/means.json
