+++ crime-applications-adaptor/src/main/resources/schemas/criminalapplicationsdatastore/maat_application_external.json
@@ -90,6 +90,6 @@
   "definitions": {
     "provider": { "$ref": "general/provider.json" },
     "address": { "$ref": "general/address.json" },
-    "means": { "$ref": "maat/means.json" }
+    "means": { "$ref": "general/means.json"}
   }
 }
